requirements.txt
astroid==2.2.5
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
fints==2.1.1
gspread==3.1.0
httplib2==0.12.1
idna==2.8
isort==4.3.16
jmespath==0.9.4
lazy-object-proxy==1.3.1
lxml==4.3.2
mccabe==0.6.1
mt-940==4.15.0
oauth2client==4.1.3
pyasn1==0.4.5
pyasn1-modules==0.2.4
pylint==2.3.1
python-dotenv==0.10.1
requests==2.21.0
rsa==4.0
sepaxml==2.0.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
webencodings==0.5.1
wrapt==1.11.1
